Send log events to the server in the order they were created.
The previous partly-reversed order made it hard to interpret them on the server when timestamps were not chronological due to network time updates or RTC failure.

The database in SPIFFS and the queue in RAM, in that order, now form a combined FIFO queue. Entries are sent to the server directly from the RAM queue if possible, but if not, moved to the flash database after a while to preserve them in case of power loss.